refactor: update import paths for project role entities and improve error handling
- Adjusted import paths for ProjectRole in project.entity.spec.ts to align with the new bounded context structure, enhancing code organization.
- Updated import paths for ProjectRole repository components in create-project.command.spec.ts to reflect the new directory structure, improving maintainability.
- Simplified error handling in optional-session.decorator.ts by removing the error parameter in the catch block, enhancing code clarity.
