+++ apps/server/src/contexts/project/use-cases/commands/create/create-project.command.spec.ts
@@ -22,8 +22,8 @@ import { Result } from '@/libs/result';
 import { MockClock } from '@/libs/time/mock-clock';
 import { PROJECT_VALIDATION_SERVICE_PORT } from '../../ports/project-validation.service.port';
 import { ProjectValidationService } from '@/contexts/project/infrastructure/services/project-validation.service';
-import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project-role/use-cases/ports/project-role.repository.port';
-import { InMemoryProjectRoleRepository } from '@/contexts/project-role/infrastructure/repositories/mock.project-role.repository';
+import { PROJECT_ROLE_REPOSITORY_PORT } from '@/contexts/project/bounded-contexts/project-role/use-cases/ports/project-role.repository.port';
+import { InMemoryProjectRoleRepository } from '@/contexts/project/bounded-contexts/project-role/infrastructure/repositories/mock.project-role.repository';
 import { GITHUB_REPOSITORY_PORT } from '@/contexts/github/use-cases/ports/github-repository.port';
 import { Octokit } from '@octokit/rest';
 import { CATEGORY_REPOSITORY_PORT } from '@/contexts/category/use-cases/ports/category.repository.port';
